fixed for the lenet the mistake that produced different results

+++ agent/ddpg.py
@@ -152,7 +152,7 @@ class DDPG(object):
             ])
 
         target_q_batch = to_tensor(reward_batch) + \
-                         self.discount * to_tensor(terminal_batch.astype(np.float)) * next_q_values
+                         self.discount * to_tensor(terminal_batch.astype(np.float32)) * next_q_values
 
         # Critic update
         self.critic.zero_grad()
